Refactor skill registry install helpers
Split the oversized skill registry module into focused discovery,\nloading, materialization, removal, staged-install, and test helper\nsubmodules while preserving the existing public API.\n\nThis keeps registry.rs below the repository file-size limit, removes\nredundant tempdir assertions from the bundle install tests, and records\nthe structural follow-up in the execplan.
